Move the logic for consuming events into ReceiptAccumulator
